Move node replacement operator to nodes_replacement.py.
