+++ __init__.py
@@ -102,7 +102,7 @@ utils_modules = [
 ui_modules = [
     "color_def", "sv_IO_panel", "sv_templates_menu",
     "sv_panels", "nodeview_rclick_menu", "nodeview_space_menu", "nodeview_keymaps",
-    "monad", "sv_icons", "presets",
+    "monad", "sv_icons", "presets", "nodes_replacement",
     # bgl modules
     "viewer_draw", "viewer_draw_mk2", "nodeview_bgl_viewer_draw", "nodeview_bgl_viewer_draw_mk2",
     "index_viewer_draw", "bgl_callback_3dview",
